requirements.txt
pytorch
numpy
pandas 
tqdm
Pillow